Made more progress on the Floating Canvas implementation.
Added a `pick_canvas` method for determining which Canvas is directly
under a given position. This is now used to determine which canvas is
directly under the mouse once per update.

Added a `UserInput` type for collecting the user input available to a
widget depending upon whether or not each form of user input is captured
or not. This is a small first step towards #424.