Updated "canvas" example to demonstrate new Floating canvas type.
Made GlyphCache deref to internal RefCell in order to allow for public
access.

Fixed drawing order of canvasses and widgets, however it is currently a
bit crude (requires allocations and sorts every Ui::draw). Definitely
potential for optimisation here - I think we should consider moving the
Vec caches over to a petgraph::Graph (and allow the edges to describe
the layering).

A few other behavioural fixes.