Fix Panic When TextBox is Empty
Previously if all the characters were deleted from a text box then we
would try to calculate the height based on 0 lines of text. This causes
the thread to panic with an arithmetic overflow on OS X.

This commit fixes things by rendering a text box with 0 characters in it
as the height of a single line.