Remove redundant test - widgets should capture mouse when highlighted if no other widget is already captured.
Add beginnings of a `draw_glium` feature and associated module. Going to
try get this working in the `glutin_glium` example alone first.

Replace the `TextRenderer` type in the `draw_piston` module in favour of
a more general `Renderer` modules which holds all relevant data for
drawing besides the `Primitives` themselves.

Add a helper function for constructing a `RawEvent::Render` as some
windows will have no events that directly match it.