Fix `TextEdit` cursor jump bugs
Fix bugs in `TextEdit` where the cursor can jump to unexpected locations
(usually the upper left of the screen) while editing. This could happen
when pressing the left arrow key or backspace at the beginning of a line,
or when pressing space at the end of a line.

The causes were minor mistakes like off-by-one errors.