Remove the `event::Raw` type in favour of taking `Input` events.
Prior to this commit, conrod took `event::Raw` as the event type that
drove forward the `Ui` and from which the `Ui` would generate
higher-level `Event`s. It turns out we only ever really used the
`event::Raw`'s `Input` variant, so to make our API more concise I
thought we could remove the unnecessary indirection of `event::Raw`.

One problem this does raise is that on Mac OS, we use the `event::Raw`'s
`Render` variant as a hack to track the window's size changes. This is
only because glutin (winit) is unable to emit events during resizing on
Mac OS. To get around the removal of this hack, I'm planning to submit a
PR to glutin_window (used by the piston backend present in most of the
examples) which will insert `Resize` events manually.

As a result of this change, we must also now pass the initial window
dimensions to the `Ui`'s constructor.

This commit also updates the related event module docs.