In examples, prevent event loop from going idle for at least one frame after an input event.
This is necessary in the case when a widget does not update atomically in a single frame.

This enables the event loop to go idle when there is no input, while not making app implementation
more complicated.

Hopefully this is temporary until widgets either handle updates atomically, notify the event loop
that they require an update, or widgets are initialized after the events that determine the widget state,
(ie. if a button label is changed when the button is pressed, the button label is created after
the button, and set based on whether the button was pressed.)