Remove the `glutin` feature in favour of more flexible `winit` feature
The old `glutin` feature was created before the `winit` crate was in a
usable state and only provides an event conversion function for
converting `winit::Event`s to `conrod::event::Input`s. Now that `winit`
is usable (and is the backend for windowing in the most recent `glutin`
and `glium` versions) It makes much more sense to do this under a
`winit` feature.

This should allow users to use conrod in non-opengl `winit`
applications.