Use Condvar to make the EventLoop thread-safe
The event loop used to need to be mutable and was thus hard to use
when you had several threads that needed to call needs_update(). Use
std::sync::Condvar and friends to make it so the EventLoop doesn't
need to be mutable at all and handles several threads properly.

next() will now busy loop at 60FPS while waiting for input. Ideally
there would be a way to have the glium Window also fire needs_update()
calls and then we could bump the timeout to something much higher or
even just wait forever until there's an event. I don't see anything
in the glium API that lets us do that cleanly though and this busy
waiting is cheap (but may be a waste of power on mobile/laptops).