Add a simple "Hello World!" text example.
This example aims to be a simple as possible demonstration of how to use
conrod with a basic winit loop using glium to draw to the screen. The
example avoids using the `examples/support` module in order to remain
copy-paste friendly. In turn, the `EventLoop` from the
`examples/support` module has been inlined into the main function.