Add `Widget::is_over` method - allows for more precise widget picking
This adds a new API for allowing a widget to optionally specify a
function that can be used for determining whether or not a given point
is over the widget. By default, a function that simply checks the
bounding rectangle is used.

This has been implemented for all existing widgets with non-rectangular
bounds (other than `Text`) including:

- `Line`
- `Oval`
- `PointPath`
- `Polygon`
- `RoundedRectangle`
- `Triangles`

Each of the containing modules also provide helper functions that may
aid in the implementation of custom widgets.