update glutin, gfx too
The change to 'texture::Mipmap::Provided' is completely guessed.
I have no understanding of the surrounding code.
They look like shader resources, not actual graphical textures,
for which mipmaps are normally not used. Hence, assuming that the
data is 'provided' (i.e. empty and no extra work is to be done).