Fix `RangeSlider` handle issue. Add `RangeSlider::skew` method.
Fixes an issue where a range slider could occasionally get stuck if the
start of the range is equal to the end of the range.

The handle also now has a minimum visible width so that if the start and
end of the range are equal, the user can still see what the value is
(rather than the slider appearing to be invisible).

A `skew` method has been added that matches the behaviour of the
`Slider::skew` method. See the method docs for details.