Don't crash in case there's an image missing
In multi-threaded programs it's common to have one thread changing
state and another displaying. It's easy in that situation to have
an image already be missing from the map while a frame that still
depends on it is still being rendered. Fix that by just not displaying
that image. Might as well have a visual glitch instead of a crash.