Remove unnecessary features from examples. Add conrod_example_shared.
Removes use of feature flags throughout example that used to be
necessary when backends where added via features.

The conrod_example_shared crate is a local crate (not intended to be
published) in which a GUI displaying (almost) all widgets can be
instantiated to edit a basic demo app. This same GUI is re-used by the
`all_*.rs` example for each of the backends to ensure they all behave
the same.