Move conversion of `Primitives` into commands and vertices into core
This adds a new `mesh` module to `conrod_core` that aims to assist with
the conversion from `Primitives` to commands and vertices. This
includes:
- Updating the glyph cache for all text and indicating whether or not
  the pixel buffer requires being re-uploaded to the GPU.
- Generating scizzor commands when necessary.
- Separating the drawing of different images into different commands.
- Generating `Draw` commands and providing the necessary slice of
  vertices for each.

At the very least, it should be possible to use this in both the
`vulkano` and `rendy` backends. It should also be possible to use this
within the opengl-based backends, however produced vertices will require
re-mapping from the Vulkan coordinate space to the OpenGL coordinate
space.